refactor(onboard): extract extra-provider reconciliation into its own module
PR Review Advisor flagged sandbox-provider-cleanup.ts monolith growth
(PRA-2, required): reconciliation is a fourth responsibility next to
detach, pre-delete cleanup, and delete recovery. Move
reconcileRegisteredExtraProviders to
src/lib/onboard/extra-provider-reconciliation.ts with a re-export from
the cleanup module so the onboard entrypoint import stays net-zero;
the cleanup module shrinks below its pre-PR line count.

Also folds in the advisor's hardening asks: SOURCE_OF_TRUTH_REVIEW
block documenting the drift's origin and removal condition (PRA-6),
a debug note when the fail-open path keeps a record after a
non-not-found probe failure (PRA-4), and a forget() guard so a
registry write failure cannot abort onboarding (PRA-5). Unit tests
move to test/extra-provider-reconciliation.test.ts alongside the
module.
